+++ src/main/java/com/example/web/controller/admin/MypageController.java
@@ -3,10 +3,10 @@ package com.example.web.controller.admin;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.util.Map;
 
 import javax.validation.Valid;
 
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -17,11 +17,15 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.multipart.MultipartFile;
 
 import com.example.dto.EmployeeDetail;
+import com.example.security.AuthenticatedUser;
+import com.example.security.vo.LoginUser;
 import com.example.service.admin.EmployeeService;
+import com.example.vo.Employee;
 import com.example.web.request.EmployeeModifyForm;
 
 @Controller
@@ -36,22 +40,20 @@ public class MypageController {
 	
 	// 내 정보 조회
 	@GetMapping("/mypage")
-	public String mypage(@RequestParam(name = "empId") String empId, Model model) {
-		EmployeeDetail employeeDetail = employeeService.getEmployeeDetail(empId);
+	public String mypage(@AuthenticatedUser LoginUser loginUser, Model model) {
+		EmployeeDetail employeeDetail = employeeService.getEmployeeDetail(loginUser.getId());
 		model.addAttribute("employee", employeeDetail);
 		
-		return "admin/employee/mypage";
+		return "/employee/mypage";
 	}
 	
 	// 내 정보 수정페이지
 	@GetMapping("/modify")
-	public String mypageModifyForm(@RequestParam(name = "empId") String empId, Model model) {
-		EmployeeDetail employeeDetail = employeeService.getEmployeeDetail(empId);
-		EmployeeModifyForm form = new EmployeeModifyForm();
-		BeanUtils.copyProperties(employeeDetail, form);
-		model.addAttribute("form", form);
+	public String modifyForm(@AuthenticatedUser LoginUser loginUser, Model model) {
+		EmployeeDetail employeeDetail = employeeService.getEmployeeDetail(loginUser.getId());
+		model.addAttribute("employee", employeeDetail);
 		
-		return "admin/employee/mypage-modify";
+		return "/employee/modify";
 	}
 	
 	// 내 정보 수정
